refactor 03
